#76: renamed files in search directories, so avoid this 'index.html,index.html,index.html' problem
