+++ assets/js/search/publications/publication-facets.js
