+++ assets/js/search/publications/publication-query.js
